requirements.txt
langchain_groq
langchain_community
langchain_core
langgraph
pypdf
python-dotenv
streamlit
pillow
